ensureChunk to RuntimeModule
step 1: non HMR cases
