+++ lib/APIPlugin.js
@@ -61,8 +61,7 @@ const REPLACEMENTS = {
 	},
 	"require.onError": {
 		expr: RuntimeGlobals.uncaughtErrorHandler,
-		req: [RuntimeGlobals.uncaughtErrorHandler],
-		type: "function"
+		req: [RuntimeGlobals.uncaughtErrorHandler]
 	}
 };
 /* eslint-enable camelcase */
