+++ lib/ChunkGraph.js
@@ -18,6 +18,7 @@ const {
 /** @typedef {import("./ChunkGroup")} ChunkGroup */
 /** @typedef {import("./Module")} Module */
 /** @typedef {import("./ModuleGraph")} ModuleGraph */
+/** @typedef {import("./RuntimeModule")} RuntimeModule */
 
 const compareModuleIterables = compareIterables(compareModulesByIdentifier);
 
@@ -113,10 +114,12 @@ class ChunkGraphChunk {
 		this.modules = new SortableSet();
 		/** @type {Map<Module, ChunkGroup>} */
 		this.entryModules = new Map();
-		/** @type {SortableSet<Module>} */
+		/** @type {SortableSet<RuntimeModule>} */
 		this.runtimeModules = new SortableSet();
 		/** @type {Set<string>} */
 		this.runtimeRequirements = new Set();
+		/** @type {Set<string>} */
+		this.runtimeRequirementsInTree = new Set();
 	}
 }
 
@@ -251,8 +254,8 @@ class ChunkGraph {
 
 		for (const chunk of oldCgm.runtimeInChunks) {
 			const cgc = this._getChunkGraphChunk(chunk);
-			cgc.runtimeModules.delete(oldModule);
-			cgc.runtimeModules.add(newModule);
+			cgc.runtimeModules.delete(/** @type {RuntimeModule} */ (oldModule));
+			cgc.runtimeModules.add(/** @type {RuntimeModule} */ (newModule));
 			newCgm.runtimeInChunks.add(chunk);
 		}
 		oldCgm.runtimeInChunks.clear();
@@ -661,7 +664,7 @@ class ChunkGraph {
 
 	/**
 	 * @param {Chunk} chunk the new chunk
-	 * @param {Module} module the runtime module
+	 * @param {RuntimeModule} module the runtime module
 	 * @returns {void}
 	 */
 	connectChunkAndRuntimeModule(chunk, module) {
@@ -685,7 +688,7 @@ class ChunkGraph {
 
 	/**
 	 * @param {Chunk} chunk the new chunk
-	 * @param {Module} module the entry module
+	 * @param {RuntimeModule} module the runtime module
 	 * @returns {void}
 	 */
 	disconnectChunkAndRuntimeModule(chunk, module) {
@@ -750,7 +753,7 @@ class ChunkGraph {
 
 	/**
 	 * @param {Chunk} chunk the chunk
-	 * @returns {Iterable<Module>} iterable of modules (do not modify)
+	 * @returns {Iterable<RuntimeModule>} iterable of modules (do not modify)
 	 */
 	getChunkRuntimeModulesIterable(chunk) {
 		const cgc = this._getChunkGraphChunk(chunk);
@@ -870,6 +873,17 @@ class ChunkGraph {
 	}
 
 	/**
+	 * @param {Chunk} chunk the chunk
+	 * @param {Iterable<string>} items runtime requirements to be added
+	 * @returns {void}
+	 */
+	addTreeRuntimeRequirements(chunk, items) {
+		const cgc = this._getChunkGraphChunk(chunk);
+		const runtimeRequirements = cgc.runtimeRequirementsInTree;
+		for (const item of items) runtimeRequirements.add(item);
+	}
+
+	/**
 	 * @param {Module} module the module
 	 * @returns {Set<string>} runtime requirements
 	 */
@@ -885,6 +899,15 @@ class ChunkGraph {
 	getChunkRuntimeRequirements(chunk) {
 		const cgc = this._getChunkGraphChunk(chunk);
 		return cgc.runtimeRequirements;
+	}
+
+	/**
+	 * @param {Chunk} chunk the chunk
+	 * @returns {Set<string>} runtime requirements
+	 */
+	getTreeRuntimeRequirements(chunk) {
+		const cgc = this._getChunkGraphChunk(chunk);
+		return cgc.runtimeRequirementsInTree;
 	}
 
 	// TODO remove in webpack 6
