+++ lib/Compilation.js
@@ -1391,6 +1391,8 @@ class Compilation {
 			for (const r of set) {
 				this.hooks.runtimeRequirementInTree.for(r).call(chunk, set);
 			}
+
+			chunkGraph.addTreeRuntimeRequirements(chunk, set);
 		}
 	}
