+++ lib/MainTemplate.js
@@ -78,12 +78,6 @@ module.exports = class MainTemplate {
 				"moduleIdExpression",
 				"renderContext"
 			]),
-			/** @type {SyncWaterfallHook<string, string, RenderBootstrapContext>} */
-			requireEnsure: new SyncWaterfallHook([
-				"source",
-				"chunkIdExpression",
-				"renderContext"
-			]),
 			/** @type {SyncWaterfallHook<string, RenderBootstrapContext>} */
 			bootstrap: new SyncWaterfallHook(["source", "renderContext"]),
 			/** @type {SyncWaterfallHook<string, Chunk, string>} */
@@ -131,10 +125,20 @@ module.exports = class MainTemplate {
 		this.hooks.startup.tap(
 			"MainTemplate",
 			(source, { chunk, hash, chunkGraph }) => {
+				if (
+					chunkGraph
+						.getTreeRuntimeRequirements(chunk)
+						.has(RuntimeGlobals.startup)
+				) {
+					return Template.asString([
+						"// run modules when ready",
+						`return ${RuntimeGlobals.startup}();`
+					]);
+				}
 				/** @type {string[]} */
-				const buf = [];
 				if (chunkGraph.getNumberOfEntryModules(chunk) > 0) {
-					const runtimeRequirements = chunkGraph.getChunkRuntimeRequirements(
+					const buf = [];
+					const runtimeRequirements = chunkGraph.getTreeRuntimeRequirements(
 						chunk
 					);
 					buf.push("// Load entry module and return exports");
@@ -158,8 +162,9 @@ module.exports = class MainTemplate {
 							)}(${moduleIdExpr});`
 						);
 					}
+					return Template.asString(buf);
 				}
-				return Template.asString(buf);
+				return source;
 			}
 		);
 		this.hooks.render.tap(
@@ -251,43 +256,8 @@ module.exports = class MainTemplate {
 			(source, renderContext) => {
 				const { chunk, chunkGraph } = renderContext;
 				const buf = [];
-				const chunkMaps = chunk.getChunkMaps();
-				// Check if there are non initial chunks which need to be imported using require-ensure
-				if (Object.keys(chunkMaps.hash).length) {
-					buf.push("// This file contains only the entry chunk.");
-					buf.push("// The chunk loading function for additional chunks");
-					buf.push(
-						`${RuntimeGlobals.ensureChunk} = function requireEnsure(chunkId) {`
-					);
-					buf.push(Template.indent("var promises = [];"));
-					buf.push(
-						Template.indent(
-							this.hooks.requireEnsure.call("", "chunkId", renderContext)
-						)
-					);
-					buf.push(Template.indent("return Promise.all(promises);"));
-					buf.push("};");
-				} else if (
-					chunkGraph.hasModuleInGraph(chunk, m =>
-						m.blocks.some(b => {
-							const chunkGroup = chunkGraph.getBlockChunkGroup(b);
-							return chunkGroup && chunkGroup.chunks.length > 0;
-						})
-					)
-				) {
-					// There async blocks in the graph, so we need to add an empty requireEnsure
-					// function anyway. This can happen with multiple entrypoints.
-					buf.push("// The chunk loading function for additional chunks");
-					buf.push("// Since all referenced chunks are already included");
-					buf.push("// in this file, this function is empty here.");
-					buf.push(
-						`${RuntimeGlobals.ensureChunk} = function requireEnsure() {`
-					);
-					buf.push(Template.indent("return Promise.resolve();"));
-					buf.push("};");
-				}
 
-				const runtimeRequirements = chunkGraph.getChunkRuntimeRequirements(
+				const runtimeRequirements = chunkGraph.getTreeRuntimeRequirements(
 					chunk
 				);
