+++ lib/RuntimeGlobals.js
@@ -46,6 +46,11 @@ exports.moduleFactories = "__webpack_require__.m";
 exports.ensureChunk = "__webpack_require__.e";
 
 /**
+ * the chunk ensure function
+ */
+exports.ensureChunkHandlers = "__webpack_require__.f";
+
+/**
  * the exported property define getter function
  */
 exports.definePropertyGetter = "__webpack_require__.d";
@@ -94,3 +99,8 @@ exports.chunkName = "__webpack_require__.cn";
  * the filename of the script part of the chunk
  */
 exports.getChunkScriptFilename = "__webpack_require__.u";
+
+/**
+ * startup signal from runtime
+ */
+exports.startup = "__webpack_require__.x";
