+++ lib/RuntimeModule.js
@@ -22,9 +22,10 @@ const Module = require("./Module");
 /** @typedef {SourceContext} GenerateContext */
 
 class RuntimeModule extends Module {
-	constructor(name) {
+	constructor(name, stage = 0) {
 		super("runtime");
 		this.name = name;
+		this.stage = stage;
 		this.buildMeta = {};
 		this.buildInfo = {};
 		this._cachedGeneratedCode = undefined;
