+++ lib/RuntimePlugin.js
@@ -10,10 +10,15 @@ const Template = require("./Template");
 const CompatGetDefaultExportRuntimeModule = require("./runtime/CompatGetDefaultExportRuntimeModule");
 const CreateFakeNamespaceObjectRuntimeModule = require("./runtime/CreateFakeNamespaceObjectRuntimeModule");
 const DefinePropertyGetterRuntimeModule = require("./runtime/DefinePropertyGetterRuntimeModule");
+const EnsureChunkRuntimeModule = require("./runtime/EnsureChunkRuntimeModule");
 const GetChunkFilenameRuntimeModule = require("./runtime/GetChunkFilenameRuntimeModule");
 const MakeNamespaceObjectRuntimeModule = require("./runtime/MakeNamespaceObjectRuntimeModule");
 const PublicPathRuntimeModule = require("./runtime/PublicPathRuntimeModule");
-const { compareIds } = require("./util/comparators");
+const {
+	concatComparators,
+	compareSelect,
+	compareIds
+} = require("./util/comparators");
 
 /** @typedef {import("./Chunk")} Chunk */
 /** @typedef {import("./Compiler")} Compiler */
@@ -31,11 +36,7 @@ const DEPENDENCIES = {
 		RuntimeGlobals.makeNamespaceObject
 	],
 	[RuntimeGlobals.definePropertyGetter]: [RuntimeGlobals.require],
-	[RuntimeGlobals.ensureChunk]: [
-		RuntimeGlobals.require,
-		RuntimeGlobals.publicPath,
-		RuntimeGlobals.getChunkScriptFilename
-	],
+	[RuntimeGlobals.ensureChunk]: [RuntimeGlobals.require],
 	[RuntimeGlobals.entryModuleId]: [RuntimeGlobals.require],
 	[RuntimeGlobals.getFullHash]: [RuntimeGlobals.require],
 	[RuntimeGlobals.makeNamespaceObject]: [RuntimeGlobals.require],
@@ -124,20 +125,39 @@ class RuntimePlugin {
 					);
 					return true;
 				});
+			compilation.hooks.runtimeRequirementInTree
+				.for(RuntimeGlobals.ensureChunk)
+				.tap("RuntimePlugin", (chunk, set) => {
+					const hasAsyncChunks = chunk.hasAsyncChunks();
+					if (hasAsyncChunks) {
+						set.add(RuntimeGlobals.ensureChunkHandlers);
+					}
+					compilation.addRuntimeModule(
+						chunk,
+						new EnsureChunkRuntimeModule(hasAsyncChunks)
+					);
+					return true;
+				});
 			compilation.mainTemplate.hooks.runtime.tap(
 				"RuntimePlugin",
 				(source, { chunk, chunkGraph }) => {
 					const buf = [];
-					if (chunkGraph.getNumberOfRuntimeModules(chunk) > 0) {
+					const runtimeModules = Array.from(
+						chunkGraph.getChunkRuntimeModulesIterable(chunk)
+					);
+					if (runtimeModules.length > 0) {
 						buf.push("// Bootstrap all runtime modules");
-						const runtimeModules = Array.from(
-							chunkGraph.getChunkRuntimeModulesIterable(chunk),
-							m => chunkGraph.getModuleId(m)
+						runtimeModules.sort(
+							concatComparators(
+								compareSelect(r => r.stage, compareIds),
+								compareSelect(r => chunkGraph.getModuleId(r), compareIds)
+							)
 						);
-						runtimeModules.sort(compareIds);
-						for (const moduleId of runtimeModules) {
+						for (const module of runtimeModules) {
 							buf.push(
-								`modules[${JSON.stringify(moduleId)}](0,0,__webpack_require__);`
+								`// ${module.name}\nmodules[${JSON.stringify(
+									chunkGraph.getModuleId(module)
+								)}](0,0,__webpack_require__);`
 							);
 						}
 					}
