+++ lib/RuntimeTemplate.js
@@ -634,10 +634,6 @@ class RuntimeTemplate {
 		}
 	}
 
-	onError() {
-		return RuntimeGlobals.uncaughtErrorHandler;
-	}
-
 	/**
 	 * @param {Object} options options
 	 * @param {string} options.exportsArgument the name of the exports object
