+++ lib/WebpackOptionsApply.js
@@ -97,8 +97,8 @@ class WebpackOptionsApply extends OptionsApply {
 		compiler.dependencies = options.dependencies;
 		if (typeof options.target === "string") {
 			let JsonpTemplatePlugin;
-			let FetchCompileWasmTemplatePlugin;
-			let ReadFileCompileWasmTemplatePlugin;
+			let FetchCompileWasmPlugin;
+			let ReadFileCompileWasmPlugin;
 			let NodeSourcePlugin;
 			let NodeTargetPlugin;
 			let NodeTemplatePlugin;
@@ -106,10 +106,10 @@ class WebpackOptionsApply extends OptionsApply {
 			switch (options.target) {
 				case "web":
 					JsonpTemplatePlugin = require("./web/JsonpTemplatePlugin");
-					FetchCompileWasmTemplatePlugin = require("./web/FetchCompileWasmTemplatePlugin");
+					FetchCompileWasmPlugin = require("./web/FetchCompileWasmPlugin");
 					NodeSourcePlugin = require("./node/NodeSourcePlugin");
 					new JsonpTemplatePlugin().apply(compiler);
-					new FetchCompileWasmTemplatePlugin({
+					new FetchCompileWasmPlugin({
 						mangleImports: options.optimization.mangleWasmImports
 					}).apply(compiler);
 					new FunctionModulePlugin().apply(compiler);
@@ -118,10 +118,10 @@ class WebpackOptionsApply extends OptionsApply {
 					break;
 				case "webworker": {
 					let WebWorkerTemplatePlugin = require("./webworker/WebWorkerTemplatePlugin");
-					FetchCompileWasmTemplatePlugin = require("./web/FetchCompileWasmTemplatePlugin");
+					FetchCompileWasmPlugin = require("./web/FetchCompileWasmPlugin");
 					NodeSourcePlugin = require("./node/NodeSourcePlugin");
 					new WebWorkerTemplatePlugin().apply(compiler);
-					new FetchCompileWasmTemplatePlugin({
+					new FetchCompileWasmPlugin({
 						mangleImports: options.optimization.mangleWasmImports
 					}).apply(compiler);
 					new FunctionModulePlugin().apply(compiler);
@@ -132,12 +132,12 @@ class WebpackOptionsApply extends OptionsApply {
 				case "node":
 				case "async-node":
 					NodeTemplatePlugin = require("./node/NodeTemplatePlugin");
-					ReadFileCompileWasmTemplatePlugin = require("./node/ReadFileCompileWasmTemplatePlugin");
+					ReadFileCompileWasmPlugin = require("./node/ReadFileCompileWasmPlugin");
 					NodeTargetPlugin = require("./node/NodeTargetPlugin");
 					new NodeTemplatePlugin({
 						asyncChunkLoading: options.target === "async-node"
 					}).apply(compiler);
-					new ReadFileCompileWasmTemplatePlugin({
+					new ReadFileCompileWasmPlugin({
 						mangleImports: options.optimization.mangleWasmImports
 					}).apply(compiler);
 					new FunctionModulePlugin().apply(compiler);
@@ -192,11 +192,11 @@ class WebpackOptionsApply extends OptionsApply {
 					break;
 				case "electron-renderer":
 					JsonpTemplatePlugin = require("./web/JsonpTemplatePlugin");
-					FetchCompileWasmTemplatePlugin = require("./web/FetchCompileWasmTemplatePlugin");
+					FetchCompileWasmPlugin = require("./web/FetchCompileWasmPlugin");
 					NodeTargetPlugin = require("./node/NodeTargetPlugin");
 					ExternalsPlugin = require("./ExternalsPlugin");
 					new JsonpTemplatePlugin().apply(compiler);
-					new FetchCompileWasmTemplatePlugin({
+					new FetchCompileWasmPlugin({
 						mangleImports: options.optimization.mangleWasmImports
 					}).apply(compiler);
 					new FunctionModulePlugin().apply(compiler);
