+++ lib/dependencies/RequireEnsureDependency.js
@@ -5,6 +5,7 @@
 
 "use strict";
 
+const RuntimeGlobals = require("../RuntimeGlobals");
 const makeSerializable = require("../util/makeSerializable");
 const NullDependency = require("./NullDependency");
 
@@ -89,7 +90,9 @@ RequireEnsureDependency.Template = class RequireEnsureDependencyTemplate extends
 			source.replace(
 				contentRange[1],
 				range[1] - 1,
-				`).bind(null, __webpack_require__)).catch(${runtimeTemplate.onError()})`
+				`).bind(null, __webpack_require__)).catch(${
+					RuntimeGlobals.uncaughtErrorHandler
+				})`
 			);
 		}
 	}
