+++ lib/node/ReadFileCompileWasmPlugin.js
@@ -5,12 +5,13 @@
 
 "use strict";
 
+const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
-const WasmMainTemplatePlugin = require("../wasm/WasmMainTemplatePlugin");
+const WasmChunkLoadingRuntimePlugin = require("../wasm/WasmChunkLoadingRuntimePlugin");
 
 /** @typedef {import("../Compiler")} Compiler */
 
-class ReadFileCompileWasmTemplatePlugin {
+class ReadFileCompileWasmPlugin {
 	constructor(options) {
 		this.options = options || {};
 	}
@@ -21,7 +22,7 @@ class ReadFileCompileWasmTemplatePlugin {
 	 */
 	apply(compiler) {
 		compiler.hooks.thisCompilation.tap(
-			"ReadFileCompileWasmTemplatePlugin",
+			"ReadFileCompileWasmPlugin",
 			compilation => {
 				const generateLoadBinaryCode = path =>
 					Template.asString([
@@ -50,20 +51,30 @@ class ReadFileCompileWasmTemplatePlugin {
 						"})"
 					]);
 
-				const plugin = new WasmMainTemplatePlugin(
-					compilation,
-					Object.assign(
-						{
-							generateLoadBinaryCode,
-							supportsStreaming: false
-						},
-						this.options
-					)
-				);
-				plugin.apply(compilation.mainTemplate);
+				compilation.hooks.runtimeRequirementInTree
+					.for(RuntimeGlobals.ensureChunkHandlers)
+					.tap("ReadFileCompileWasmPlugin", (chunk, set) => {
+						const chunkGraph = compilation.chunkGraph;
+						if (
+							!chunkGraph.hasModuleInGraph(chunk, m =>
+								m.type.startsWith("webassembly")
+							)
+						) {
+							return;
+						}
+						set.add(RuntimeGlobals.moduleCache);
+						compilation.addRuntimeModule(
+							chunk,
+							new WasmChunkLoadingRuntimePlugin(chunk, compilation, {
+								generateLoadBinaryCode,
+								supportsStreaming: false,
+								mangleImports: this.options.mangleImports
+							})
+						);
+					});
 			}
 		);
 	}
 }
 
-module.exports = ReadFileCompileWasmTemplatePlugin;
+module.exports = ReadFileCompileWasmPlugin;
