+++ lib/runtime/GetChunkFilenameRuntimeModule.js
@@ -54,6 +54,7 @@ class GetChunkFilenameRuntimeModule extends RuntimeModule {
 					[contentType]: length => {
 						const shortContentHashMap = {};
 						const contentHash = chunkMaps.contentHash[contentType];
+						if (!contentHash) return "XXXX";
 						for (const chunkId of Object.keys(contentHash)) {
 							if (typeof contentHash[chunkId] === "string") {
 								shortContentHashMap[chunkId] = contentHash[chunkId].substr(
