+++ lib/webpack.js
@@ -170,8 +170,7 @@ exportPlugins((exports.optimize = {}), {
 	SplitChunksPlugin: () => require("./optimize/SplitChunksPlugin")
 });
 exportPlugins((exports.web = {}), {
-	FetchCompileWasmTemplatePlugin: () =>
-		require("./web/FetchCompileWasmTemplatePlugin"),
+	FetchCompileWasmPlugin: () => require("./web/FetchCompileWasmPlugin"),
 	JsonpTemplatePlugin: () => require("./web/JsonpTemplatePlugin")
 });
 exportPlugins((exports.webworker = {}), {
@@ -179,8 +178,7 @@ exportPlugins((exports.webworker = {}), {
 });
 exportPlugins((exports.node = {}), {
 	NodeTemplatePlugin: () => require("./node/NodeTemplatePlugin"),
-	ReadFileCompileWasmTemplatePlugin: () =>
-		require("./node/ReadFileCompileWasmTemplatePlugin")
+	ReadFileCompileWasmPlugin: () => require("./node/ReadFileCompileWasmPlugin")
 });
 exportPlugins((exports.debug = {}), {
 	ProfilingPlugin: () => require("./debug/ProfilingPlugin")
