--- test/cases/runtime/issue-2391-chunk/index.js
@@ -1,4 +1,0 @@
-it("should have a require.onError function by default", function() {
-	expect((typeof require.onError)).toBe("function");
-	require(["./file"]);
-});
