--- test/cases/runtime/issue-2391/index.js
@@ -1,3 +1,0 @@
-it("should not have a require.onError function by default", function() {
-	expect((typeof require.onError)).toBe("undefined"); // expected to fail in browsertests
-});
